GuidedTours: Move state to its own directory in state/ui
