+++ client/boot/index.js
@@ -25,7 +25,7 @@ var config = require( 'config' ),
 	user = require( 'lib/user' )(),
 	receiveUser = require( 'state/users/actions' ).receiveUser,
 	setCurrentUserId = require( 'state/current-user/actions' ).setCurrentUserId,
-	showGuidedTour = require( 'state/ui/actions' ).showGuidedTour,
+	showGuidedTour = require( 'state/ui/guided-tours/actions' ).showGuidedTour,
 	sites = require( 'lib/sites-list' )(),
 	superProps = require( 'lib/analytics/super-props' ),
 	i18n = require( 'lib/mixins/i18n' ),
