+++ client/state/ui/reducer.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import { combineReducers } from 'redux';
-import omit from 'lodash/omit';
 
 /**
  * Internal dependencies
@@ -12,10 +11,9 @@ import {
 	SET_SECTION,
 	SERIALIZE,
 	DESERIALIZE,
-	SHOW_GUIDED_TOUR,
-	UPDATE_GUIDED_TOUR,
 } from 'state/action-types';
 import editor from './editor/reducer';
+import guidedTour from './guided-tours/reducer';
 import reader from './reader/reducer';
 
 /**
@@ -75,23 +73,6 @@ export function isLoading( state = false, action ) {
 	switch ( action.type ) {
 		case SET_SECTION:
 			return ( action.isLoading !== undefined ) ? action.isLoading : state;
-	}
-	return state;
-}
-
-export function guidedTour( state = {}, action ) {
-	switch ( action.type ) {
-		case SHOW_GUIDED_TOUR:
-			const { stepName = 'init' } = action;
-			return {
-				stepName,
-				shouldShow: action.shouldShow,
-				shouldDelay: action.shouldDelay,
-				shouldReallyShow: ( action.shouldShow || state.shouldShow ) && ! action.shouldDelay,
-				tour: action.tour,
-			};
-		case UPDATE_GUIDED_TOUR:
-			return Object.assign( {}, state, omit( action, 'type' ) );
 	}
 	return state;
 }
